--- capacitor.config.ts
@@ -1,10 +1,0 @@
-import { CapacitorConfig } from '@capacitor/cli';
-
-const config: CapacitorConfig = {
-  appId: 'io.ionic.starter',
-  appName: 'wt23-foodfinder-ionic-react-app',
-  webDir: 'build',
-  bundledWebRuntime: false
-};
-
-export default config;
