+++ src/components/Maps.js
@@ -1,7 +1,8 @@
-import React from 'react'
+import React from 'react';
 import { GoogleMap, LoadScript } from '@react-google-maps/api';
-import { Marker } from '@react-google-maps/api';
-//const ScriptLoaded = require("../../docs/ScriptLoaded").default;
+import { MarkerF } from '@react-google-maps/api';
+import { GetLat, GetLng } from './GetLatLng';
+//const ScriptLoaded = require("./ScriptLoaded").default;
 
 
 const containerStyle = {
@@ -11,67 +12,47 @@ const containerStyle = {
 
 const googleMapsApiKey = process.env.REACT_APP_GOOGLE_MAPS_API_KEY;
 
+// Maps
+
+function Maps() {
+  
+ const onLoad = marker => {
+    console.log('marker: ', marker)
+  }
 // example marker position
 
-const markerPosition = {
-  lat: 52.489587,
-  lng: 13.485512
-}
+  const markerPosition = {
+    lat: 52.489587,
+    lng: 13.485512
+  }
 
-const markerPosition2 = {
-  lat: 52.490766,
-  lng: 13.483752
-}
+  // user's location
 
-const onLoad = marker => {
-  console.log('marker: ', marker)
-}
-
-// class component Maps
-
-class Maps extends React.Component {
-  
-  constructor(props){
-    super(props);
-
-    this.state = { lng: null, lat: null };
-}
-
-// get user's location
-    componentDidMount(){
-      
-      window.navigator.geolocation.getCurrentPosition(
-          (position) => this.setState({lat: position.coords.latitude, lng:position.coords.longitude}),
-
-      );        
-    }
-    render () {
+  const center = {
+    lat: GetLat(),
+    lng: GetLng()
+  }
+    
   return (
     <LoadScript
       googleMapsApiKey=  {googleMapsApiKey}
     >
       <GoogleMap
         mapContainerStyle={containerStyle}
-        center={this.state}
+        center={center}
         zoom={14}
       >
         { /* Child components, such as markers, info windows, etc. */ }
 
-      <Marker
+      <MarkerF
       onLoad={onLoad}
       position={markerPosition}
-  />
-
-<Marker
-      onLoad={onLoad}
-      position={markerPosition2}
   />
         
         <></>
       </GoogleMap>
     </LoadScript>
   )
-}
 }
