+++ src/components/MaxDistance.js
@@ -5,7 +5,8 @@ import { IonLabel, IonRange } from '@ionic/react';
 //https://ionicframework.com/docs/api/range
 
 function MaxDistanceSet() {
-    const [lastEmittedValue=0, setLastEmittedValue] = useState();
+    const [lastEmittedValue, setLastEmittedValue] = useState(0);
+    
     return (
       <>
         <IonRange onIonChange={({ detail }) => setLastEmittedValue(detail.value)} ticks={true} snaps={true} min={0} max={5}></IonRange>
